+++ CHANGELOG.md
@@ -8,6 +8,7 @@
 ```
 Sqlite operation failure database is locked attempted to run migration and failed. closing connection
 ```
+* Fix race condition causing data received during uploads not to be applied.
 
 ## 1.0.0-BETA28